refactor slot
